Add documentation. Rename references from run to cycle
